+++ assets/javascripts/kitten/components/grid/sticky-container/test.js
@@ -50,7 +50,11 @@ describe('<StickyContainer />', () => {
   describe('never sticky', () => {
     beforeEach(() => {
       component = renderer
-        .create(<StickyContainer isSticky="never" />)
+        .create(
+          <StickyContainer isSticky="never">
+            test
+          </StickyContainer>
+        )
         .toJSON()
     })