+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: `StickyContainer`: Add `"never"` value for `isSticky` prop.
+
 ## [3.21.0] - 2021-06-21
 
 Feature: